pyproject.toml
[tool.poetry]
name = "tour-guidance"
version = "0.1.0"
description = ""
authors = ["minamotooRin <minamotoorin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
geopy = "^2.4.1"
requests = "^2.32.3"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
